requirements.txt
# Core ML and Deep Learning
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress Bars and Utilities
tqdm>=4.62.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Model Serialization
joblib>=1.1.0

# Optional GPU Support (comment out if not using GPU)
# tensorflow-gpu>=2.12.0

# Development and Testing (optional)
pytest>=6.2.0
flake8>=3.9.0
black>=21.6.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"